Added modes and addons directory, cfg file creation, fixes

+++ scripts-740/instance.sh
@@ -8,7 +8,9 @@
 mkdir -p "$INSTANCE_DIR/csgo"
 
 # Directories that are fully copied, the instance owner can do whatever he wants there
-cp --reflink=auto -R "$INSTALL_DIR/csgo/addons" "$INSTANCE_DIR/csgo/addons"
+if [[ -e "$INSTALL_DIR/csgo/addons" ]]; then
+	cp --reflink=auto -R "$INSTALL_DIR/csgo/addons" "$INSTANCE_DIR/csgo/addons"
+	fi
 cp --reflink=auto -R "$INSTALL_DIR/csgo/cfg" "$INSTANCE_DIR/csgo/cfg"
 cp --reflink=auto -R "$INSTALL_DIR/csgo/models" "$INSTANCE_DIR/csgo/models"
 cp --reflink=auto -R "$INSTALL_DIR/csgo/sound" "$INSTANCE_DIR/csgo/sound"
@@ -22,3 +24,6 @@ mkdir -p "$INSTANCE_DIR/csgo/maps"
 mkdir -p "$INSTANCE_DIR/csgo/maps/cfg"
 mkdir -p "$INSTANCE_DIR/csgo/maps/soundcache"
 mkdir -p "$INSTANCE_DIR/csgo/resource/overviews"
+
+
+# Other existing files/directories are linked during the next step of the instance creation process
